Change the file format of the saved workflow. It's still XML but uses the namespace urn:qanboard:workflow:2.0 instead of urn:quanboard:workflow:1.0.
The previous format can still be read, and will be converted automatically.

The main new feature is that the tasks are saved in a separate element from the queues. This means that a task can exist even if it is not listed in any queue, meaning it has been archived. This will be useful for the planned snapshot feature that will allow to store several vues of the board at different times.